pyproject.toml
[project]
name = "hackathon-devops-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "checkov>=3.2.489",
    "fastapi==0.104.1",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "uvicorn[standard]==0.24.0",
]

[dependency-groups]
test = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]
